Merge from trunk to clang_3.5 branch r586: ------------------------------------------------------------------------ r586 | kim.grasman@gmail.com | 2014-10-05 21:38:22 +0200 (Sun, 05 Oct 2014) | 9 lines
Fix issue 140: pragma keep doesn't keep duplicates

The removal of duplicates is now countered by recording the included FileEntry
of all include lines marked "IWYU pragma: keep", and then marking all includes
of said FileEntry as desired.

We think that if the user goes to the trouble of marking an include line with
the keep pragma, they can probably explicitly remove additional includes of
that header manually if they're not wanted.
------------------------------------------------------------------------