Merge from trunk to clang_3.6 branch r604: ------------------------------------------------------------------------ r604 | kim.grasman@gmail.com | 2015-03-02 11:54:02 -0800 (Mon, 02 Mar 2015) | 12 lines
Rename GetFileEntry(FullSourceLoc) GetLocFileEntry

The use of FullSourceLoc in IWYU was confusing and error prone.
GetFileEntry was already templated on all AST node types and overloaded on
SourceLocation, so calling GetFileEntry with a FullSourceLoc would have a
completely different meaning than calling it with a plain SourceLocation.

Add comments to make it clearer that GetLocFileEntry does not do any magic.

Also inline a few pointless utility methods.

No functional change.
------------------------------------------------------------------------